major change: now using warp matrix within state estimation
this is critical because it reduces drift while doing state estimation. previously, the angle estimate from optical flow was worsened because of drift, leading to a positive feedback loop (with a negative outcome). now the drift is reduced, which puts less burden on TimTrack. you can get away with a relatively small angle gain in most cases

other changes:
- added a gain for the horizontal drift, can/should be smaller than length gain
- padded the image with grey space on left and right, to allow visualizing extrapolation